Refactor C# projects
