+++ csharp.xUnit/global.json
