--- csharp/ApprovalTest.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Text;
-using ApprovalTests;
-using ApprovalTests.Reporters;
-using NUnit.Framework;
-
-namespace csharp
-{
-    [UseReporter(typeof(DiffReporter))]
-    [TestFixture]
-    public class ApprovalTest
-    {
-        [Test]
-        public void ThirtyDays()
-        {
-            
-            StringBuilder fakeoutput = new StringBuilder();
-            Console.SetOut(new StringWriter(fakeoutput));
-            Console.SetIn(new StringReader("a\n"));
-
-            Program.Main(new string[] { "30" });
-            var output = fakeoutput.ToString();
-
-            Approvals.Verify(output);
-        }
-    }
-}
