--- csharp/GildedRoseTest.cs
@@ -1,18 +1,0 @@
-﻿using NUnit.Framework;
-using System.Collections.Generic;
-
-namespace csharp
-{
-    [TestFixture]
-    public class GildedRoseTest
-    {
-        [Test]
-        public void foo()
-        {
-            IList<Item> Items = new List<Item> { new Item { Name = "foo", SellIn = 0, Quality = 0 } };
-            GildedRose app = new GildedRose(Items);
-            app.UpdateQuality();
-            Assert.AreEqual("fixme", Items[0].Name);
-        }
-    }
-}
